sds-go: expose secondary validators configuration in the ExtraConfig
Also, do not use a pointer to the string but the string itself. If the string is
empty, the Go JSON marshaling will automatically discard this field from the final
output thanks to the `omitempty` flag, and let directly use the SecondaryValitor
consts in callers.

+ have the ProximityKeywordsConfig public in case someone wants to use it
without the `New*Rule` helpers.