Update slab 0.4.11 dependency version
Following a cargo audit, it turned out that `slab` in version 0.4.10 has a documented vulnerability. It is used as indirect dependency. This PR remediates this issue by forcing the increase of the `slab` version. unfortunately, it is not possible to only bump the direct dependencies as some of them did not fix their dependency on slab 0.4.10.

#### Summary of the vulnerability

```
Crate:     slab
Version:   0.4.10
Title:     Out-of-bounds access in `get_disjoint_mut` due to incorrect bounds check
Date:      2025-08-12
ID:        RUSTSEC-2025-0047
URL:       https://rustsec.org/advisories/RUSTSEC-2025-0047
Solution:  Upgrade to >=0.4.11
Dependency tree:
slab 0.4.10
├── quinn-proto 0.11.12
│   └── quinn 0.11.8
│       └── reqwest 0.12.28
│           └── dd-sds 0.1.2
│               ├── sds-fuzz 0.1.0
│               └── dd-sds 0.1.2
├── h2 0.4.12
│   ├── reqwest 0.12.28
│   └── hyper 1.6.0
│       ├── reqwest 0.12.28
│       ├── hyper-util 0.1.16
│       │   ├── reqwest 0.12.28
│       │   ├── hyper-rustls 0.27.7
│       │   │   └── reqwest 0.12.28
│       │   └── httpmock 0.8.2
│       │       └── dd-sds 0.1.2
│       ├── hyper-rustls 0.27.7
│       └── httpmock 0.8.2
└── futures-util 0.3.31
    ├── tower-http 0.6.8
    │   └── reqwest 0.12.28
    ├── tower 0.5.2
    │   ├── tower-http 0.6.8
    │   └── reqwest 0.12.28
    ├── reqwest 0.12.28
    ├── hyper-util 0.1.16
    ├── hyper 1.6.0
    ├── httpmock 0.8.2
    ├── futures-executor 0.3.31
    │   └── futures 0.3.31
    │       └── dd-sds 0.1.2
    └── futures 0.3.31
```